orchestrator/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
Requests==2.31.0
